Calibration of the differencing noise (sigma_epsilon)

Calibration of the differencing noise (sigma_epsilon)

Usage

pm_calib_diff(params = NULL, beta = NULL, tau = NULL, level = 0.95)

Arguments

params

object pm_params

beta

real vector of beta values (accuracy level of the inference)

tau

real vector of tau values (ceiling risk level)

level

confidence level for the confidence intervall of the relative loss

Value

data.frame and pm_calib_diff object

Details

By default, the function returns the results for some beta and tau values beta in (0.05, 0.1, 0.15, 0.2, 0.25) and tau in s(0.5, 0.55, 0.6, 0.65, ..., 0.90, 0.95)

Examples

# default behaviour
pm_calib_diff()
#>    beta  tau  sigma_eps        EZ        CI
#> 1  0.05 0.50 0.07413011  5.914727 14.529235
#> 2  0.05 0.55 0.06618878  5.281101 12.972763
#> 3  0.05 0.60 0.05940915  4.740164 11.643979
#> 4  0.05 0.65 0.05349944  4.268637 10.485697
#> 5  0.05 0.70 0.04824237  3.849184  9.455330
#> 6  0.05 0.75 0.04346506  3.468010  8.518994
#> 7  0.05 0.80 0.03901521  3.112963  7.646840
#> 8  0.05 0.85 0.03473352  2.771334  6.807645
#> 9  0.05 0.90 0.03039784  2.425397  5.957867
#> 10 0.05 0.95 0.02551067  2.035457  5.000000
#> 11 0.10 0.50 0.14826022 11.829454 29.058470
#> 12 0.10 0.55 0.13237756 10.562201 25.945526
#> 13 0.10 0.60 0.11881829  9.480328 23.287958
#> 14 0.10 0.65 0.10699887  8.537275 20.971394
#> 15 0.10 0.70 0.09648473  7.698368 18.910660
#> 16 0.10 0.75 0.08693011  6.936019 17.037989
#> 17 0.10 0.80 0.07803041  6.225926 15.293680
#> 18 0.10 0.85 0.06946705  5.542668 13.615291
#> 19 0.10 0.90 0.06079568  4.850794 11.915735
#> 20 0.10 0.95 0.05102135  4.070914 10.000000
#> 21 0.15 0.50 0.22239033 17.744181 43.587704
#> 22 0.15 0.55 0.19856634 15.843302 38.918288
#> 23 0.15 0.60 0.17822744 14.220492 34.931937
#> 24 0.15 0.65 0.16049831 12.805912 31.457090
#> 25 0.15 0.70 0.14472710 11.547552 28.365991
#> 26 0.15 0.75 0.13039517 10.404029 25.556983
#> 27 0.15 0.80 0.11704562  9.338889 22.940520
#> 28 0.15 0.85 0.10420057  8.314003 20.422936
#> 29 0.15 0.90 0.09119352  7.276191 17.873602
#> 30 0.15 0.95 0.07653202  6.106372 15.000000
#> 31 0.20 0.50 0.29652044 23.658908 58.116939
#> 32 0.20 0.55 0.26475513 21.124403 51.891051
#> 33 0.20 0.60 0.23763659 18.960657 46.575916
#> 34 0.20 0.65 0.21399774 17.074550 41.942787
#> 35 0.20 0.70 0.19296947 15.396736 37.821321
#> 36 0.20 0.75 0.17386022 13.872039 34.075978
#> 37 0.20 0.80 0.15606083 12.451853 30.587360
#> 38 0.20 0.85 0.13893409 11.085337 27.230582
#> 39 0.20 0.90 0.12159137  9.701587 23.831470
#> 40 0.20 0.95 0.10204269  8.141829 20.000000
#> 41 0.25 0.50 0.37065055 29.573635 72.646174
#> 42 0.25 0.55 0.33094391 26.405503 64.863814
#> 43 0.25 0.60 0.29704574 23.700821 58.219895
#> 44 0.25 0.65 0.26749718 21.343187 52.428484
#> 45 0.25 0.70 0.24121184 19.245920 47.276651
#> 46 0.25 0.75 0.21732528 17.340048 42.594972
#> 47 0.25 0.80 0.19507604 15.564816 38.234201
#> 48 0.25 0.85 0.17366762 13.856671 34.038227
#> 49 0.25 0.90 0.15198921 12.126984 29.789337
#> 50 0.25 0.95 0.12755336 10.177286 25.000000